Various bug fixes and Update syntax tests (#217)
* Bug fix: Conversion of literal to address was failing due to not testing for
unbound integer type.

* Bug fixes: Handled conversion of FixedBytes to Address and vice versa
in appendTypeConversion of IeleCompiler.

* Bug fix: In  conversion of Address to Integer, due to missing check
for unbound integer types.

* Implementation of .address for external functions in IeleCompiler.cpp.

* Updated failing syntax tests in types, constants, controlFlow,
functionTypes, functionCalls, inlineAssembly.

* Updated failing compilation tests.

* Updated failing execution tests.

+++ test/libsolidity/syntaxTests/constants/cyclic_dependency_2.sol
@@ -1,10 +1,10 @@
 contract C {
-    uint constant a = b * c;
-    uint constant b = 7;
-    uint constant c = b + uint(keccak256(abi.encodePacked(d)));
-    uint constant d = 2 + a;
+    uint256 constant a = b * c;
+    uint256 constant b = 7;
+    uint256 constant c = b + uint256(keccak256(abi.encodePacked(d)));
+    uint256 constant d = 2 + a;
 }
 // ----
-// TypeError 6161: (17-40): The value of the constant a has a cyclic dependency via c.
-// TypeError 6161: (71-129): The value of the constant c has a cyclic dependency via d.
-// TypeError 6161: (135-158): The value of the constant d has a cyclic dependency via a.
+// TypeError 6161: (17-43): The value of the constant a has a cyclic dependency via c.
+// TypeError 6161: (77-141): The value of the constant c has a cyclic dependency via d.
+// TypeError 6161: (147-173): The value of the constant d has a cyclic dependency via a.
